refactor(scheduler): extract _setup_jit_cache method and move dispatch_scheduler_event_loop
